backend/requirements.txt
channels==4.2.0
channels_redis==4.2.1
django-redis==5.4.0
Django==4.2.20
django-allauth==65.3.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
daphne==4.1.2
python-dotenv==1.0.1
gunicorn==23.0.0
sentry-sdk==2.22.0
prometheus-client==0.14.1 # prometheus use to custom metrics, 우선은 사용 안함
django-prometheus==2.3.1 # prometheus use to be set middleware
django-debug-toolbar==5.1.0 # django debug toolbar